+++ src/LocalDateTime.php
@@ -208,14 +208,6 @@ final class LocalDateTime implements JsonSerializable, Stringable
     }
 
     /**
-     * @deprecated Use getDayOfMonth() instead.
-     */
-    public function getDay(): int
-    {
-        return $this->date->getDayOfMonth();
-    }
-
-    /**
      * @return int<1, 31>
      */
     public function getDayOfMonth(): int