+++ src/ZonedDateTime.php
@@ -254,14 +254,6 @@ class ZonedDateTime implements JsonSerializable, Stringable
     }
 
     /**
-     * @deprecated Use getDayOfMonth() instead.
-     */
-    public function getDay(): int
-    {
-        return $this->localDateTime->getDayOfMonth();
-    }
-
-    /**
      * @return int<1, 31>
      */
     public function getDayOfMonth(): int